新增学生与班级的所有接口

+++ web-management/src/main/java/com/chen/service/impl/DeptServiceImpl.java
@@ -1,7 +1,10 @@
 package com.chen.service.impl;
 
+import com.chen.exception.DeptException;
 import com.chen.mapper.DeptMapper;
+import com.chen.mapper.EmpMapper;
 import com.chen.pojo.Dept;
+import com.chen.pojo.Emp;
 import com.chen.service.DeptService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +16,8 @@ import java.util.List;
 public class DeptServiceImpl implements DeptService {
     @Autowired
     private DeptMapper deptMapper;
+    @Autowired
+    private EmpMapper empMapper;
 
     @Override
     public List<Dept> findAll() {
@@ -20,7 +25,12 @@ public class DeptServiceImpl implements DeptService {
     }
 
     @Override
-    public void deleteById(Integer id) {
+    public void deleteById(Integer id) throws DeptException {
+        //判断该部门下是否有员工
+        List<Emp> emp=empMapper.getByDeptId(id);
+        if(emp.size()>0){
+            throw new DeptException("对不起，当前部门下有员工，不能直接删除！对不起，当前部门下有员工，不能直接删除！");
+        }
         deptMapper.deleteById(id);
     }
